feat: Implement AI Chatbot for onboarding assistance
- Added Chatbot component to facilitate user interactions and provide onboarding support.
- Integrated axios for API calls to backend for chat responses.
- Implemented message handling and scrolling behavior for chat interface.
- Created ITTicketModal for raising IT tickets with a structured form.
- Developed ManagerCard to display current user's manager information.
- Added Modal component for reusable modal functionality across the app.
- Designed Sidebar for navigation with links to various sections.
- Built Dashboard page to summarize onboarding tasks and quick links.
- Created OnboardingChecklistCard to track onboarding progress.
- Developed QuickLinksCard for easy access to important links.
- Implemented DocumentHub for viewing and accessing company documents.
- Created OnboardingChecklistPage for detailed task management.
- Established SettingsPage as a placeholder for future settings management.
- Developed TeamDirectory to display team members' information.
- Added LloydsIcon SVG asset for branding.
- Introduced mockData for team members, documents, onboarding tasks, and current user information.
