+++ src/Pages/SettingsPage.jsx
@@ -1,0 +1,10 @@
+const SettingsPage = () => (
+    <div>
+        <h1 className="text-3xl font-bold mb-6">Settings</h1>
+        <div className="bg-white dark:bg-slate-800/50 p-6 rounded-lg shadow-sm">
+            <p>Settings page is under construction. Here you will be able to manage your profile and notification preferences.</p>
+        </div>
+    </div>
+);
+
+export default SettingsPage
