+++ src/mockData.js
@@ -1,0 +1,39 @@
+
+export const teamMembers = [
+  { id: 1, name: 'Jane Smith', role: 'Engineering Manager', email: 'jane.s@example.com', phone: '123-456-7890' },
+  { id: 2, name: 'Bob Johnson', role: 'Senior Software Engineer', email: 'bob.j@example.com', phone: '123-456-7891' },
+  { id: 3, name: 'Charlie Brown', role: 'Product Manager', email: 'charlie.b@example.com', phone: '123-456-7892' },
+  { id: 4, name: 'Diana Prince', role: 'UX/UI Designer', email: 'diana.p@example.com', phone: '123-456-7893' },
+];
+
+export const documents = [
+  { id: 1, name: 'Employee Handbook', category: 'HR', url: '#' },
+  { id: 2, name: 'IT Security Policy', category: 'IT', url: '#' },
+  { id: 3, name: 'Project Phoenix Architecture Overview', category: 'Project', url: '#' },
+  { id: 4, name: '2025 Holiday Calendar', category: 'HR', url: '#' },
+  { id: 5, name: 'Expense Reimbursement Guide', category: 'Finance', url: '#' },
+];
+
+export const onboardingTasks = [
+  { id: 1, text: 'Complete HR paperwork in Workday', completed: true, day: 1 },
+  { id: 2, text: 'Set up your development environment', completed: true, day: 1 },
+  { id: 3, text: 'Schedule a 1:1 introduction with your manager', completed: false, day: 2 },
+  { id: 4, text: 'Attend company orientation session', completed: false, day: 2 },
+  { id: 5, text: 'Review the team\'s project documentation', completed: false, day: 3 },
+];
+
+export const currentUser = {
+  name: 'Alex',
+  role: 'Software Engineer',
+  manager: 'Jane Smith',
+  managerInitials: 'JS',
+  team: 'Phoenix Team',
+  joinDate: 'July 22, 2025',
+};
+
+export const quickLinks = [
+    {id: 1, name: 'Company Directory', url: '#'},
+    {id: 2, name: 'IT Helpdesk Portal', url: '#'},
+    {id: 3, name: 'Performance Reviews', url: '#'},
+]
+
